Extract serde part of xpallet-support into xp-rpc
