+++ primitives/rpc/Cargo.toml
@@ -5,4 +5,9 @@ authors = ["The ChainX Authors"]
 edition = "2018"
 
 [dependencies]
+hex = "0.4"
 jsonrpc-core = "15.0.0"
+serde = { version = "1.0", features = ["derive"] }
+
+[dev-dependencies]
+serde_json = "1.0"
