+++ xpallets/assets-registrar/Cargo.toml
@@ -21,6 +21,7 @@ frame-system = { version = "2.0.0", default-features = false }
 chainx-primitives = { path = "../../primitives", default-features = false }
 xp-assets-registrar = { path = "../../primitives/assets-registrar", default-features = false }
 xp-protocol = { path = "../../primitives/protocol", default-features = false }
+xp-rpc = { path = "../../primitives/rpc", optional = true }
 xp-runtime = { path = "../../primitives/runtime", default-features = false }
 
 # ChainX pallets
@@ -46,6 +47,7 @@ std = [
     "chainx-primitives/std",
     "xp-assets-registrar/std",
     "xp-protocol/std",
+    "xp-rpc",
     "xp-runtime/std",
     # ChainX pallets
     "xpallet-support/std",
