+++ xpallets/gateway/common/Cargo.toml
@@ -24,6 +24,7 @@ frame-system = { version = "2.0.0", default-features = false }
 chainx-primitives = { path = "../../../primitives", default-features = false }
 xp-io = { path = "../../../primitives/io", default-features = false, optional = true }
 xp-protocol = { path = "../../../primitives/protocol", default-features = false }
+xp-rpc = { path = "../../../primitives/rpc", optional = true }
 xp-runtime = { path = "../../../primitives/runtime", default-features = false }
 
 # ChainX pallets
@@ -60,6 +61,7 @@ std = [
     "chainx-primitives/std",
     "xp-io/std",
     "xp-protocol/std",
+    "xp-rpc",
     "xp-runtime/std",
     # ChainX pallets
     "xpallet-assets/std",
