+++ xpallets/mining/asset/rpc/Cargo.toml
@@ -21,8 +21,5 @@ sp-runtime = "2.0.0"
 # ChainX primitives
 xp-rpc = { path = "../../../../primitives/rpc" }
 
-# ChainX pallets
-xpallet-support = { path = "../../../support" }
-
 # ChainX pallets api
 xpallet-mining-asset-rpc-runtime-api = { path = "./runtime-api" }
