+++ xpallets/mining/asset/rpc/src/lib.rs
@@ -14,9 +14,7 @@ use sp_api::ProvideRuntimeApi;
 use sp_blockchain::HeaderBackend;
 use sp_runtime::{generic::BlockId, traits::Block as BlockT};
 
-use xp_rpc::{runtime_error_into_rpc_err, Result};
-
-use xpallet_support::{RpcBalance, RpcMiningWeight};
+use xp_rpc::{runtime_error_into_rpc_err, Result, RpcBalance, RpcMiningWeight};
 
 use xpallet_mining_asset_rpc_runtime_api::{
     AssetId, AssetLedger, MinerLedger, MiningAssetInfo, XMiningAssetApi as XMiningAssetRuntimeApi,
