+++ xpallets/mining/staking/src/types.rs
@@ -106,7 +106,7 @@ pub struct ValidatorProfile<BlockNumber> {
     /// Block number of last performed `chill` operation.
     pub last_chilled: Option<BlockNumber>,
     /// Referral identity that belongs to the validator.
-    #[cfg_attr(feature = "std", serde(with = "xpallet_support::serde_text"))]
+    #[cfg_attr(feature = "std", serde(with = "xp_rpc::serde_text"))]
     pub referral_id: ReferralId,
 }
