+++ xpallets/support/Cargo.toml
@@ -5,25 +5,17 @@ authors = ["The ChainX Authors"]
 edition = "2018"
 
 [dependencies]
-codec = { package = "parity-scale-codec", version = "1.3.4", features = ["derive"], default-features = false }
 hex = { version = "0.4", default-features = false }
-serde = { version = "1.0", optional = true }
-
 # Substrate primitives
 sp-std = { version = "2.0.0", default-features = false }
 
 # Substrate pallets
 frame-support = { version = "2.0.0", default-features = false }
 
-[dev-dependencies]
-serde_json = "1.0"
-
 [features]
 default = ["std"]
 std = [
-    "codec/std",
     "hex/std",
-    "serde",
     # Substrate primitives
     "sp-std/std",
     # Substrate pallets
